backend/seaweed-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.5.0
numpy==1.26.4
joblib==1.4.2
google-generativeai==0.7.2